requirements.txt
gradio==5.38.0
moviepy==1.0.3
Pillow==11.3.0
